Remove previous useBaseColors store & move types
